Implemented StepUntilStuck() which runs instructions until the PC is unchanged.
This is intended to detect single-instruction infinite loops, which with the DCPU-16 (and my understanding of it) are impossible to get out of.

Also bug-fixed in StepInstruction(), it wasn't handling skipping (IFx) properly. Not 100% sure it does the right thing now either, to be honest. :)

It might make sense to have the IFx instructions handle skipping themselves a bit more simply (by chaining a different Thunk).